Write method to add pairs and keys
