+++ lib/enigma.rb
@@ -1,11 +1,14 @@
+require_relative 'modules/transformable'
+
 class Enigma
+  include Transformable
 
   def initialize
+    @offset_values = nil
   end
 
   def encrypt(message, key, date)
-    key_to_pairs(key)
-
+    get_info(key, date)
     encryption_info = {}
     encryption_info[:encryption] = new_message(message.downcase)
     encryption_info[:key] = key
@@ -13,14 +16,7 @@ class Enigma
     encryption_info
   end
 
-  def key_to_pairs(key)
-    keys = []
-    key = key.split("").each_cons(2) { |pair| keys << pair }
-    keys.map { |key| key.join.to_i }
-  end
-
-  def get_date_keys(date)
-    last_four = ((((date.to_i)**2).to_s)[-4..-1])
-    last_four.split("").map { |key| key.to_i }
+  def get_info(key, date)
+    @offset_values = total_offset(key_to_pairs(key), date_to_key(date))
   end
 end
